#642 Split LeaderUtility to LeaderBiography
