+++ src/main/java/org/openRealmOfStars/game/States/RealmView.java
@@ -35,7 +35,7 @@ import org.openRealmOfStars.gui.panels.BlackPanel;
 import org.openRealmOfStars.gui.panels.RaceImagePanel;
 import org.openRealmOfStars.gui.utilies.GuiStatics;
 import org.openRealmOfStars.player.PlayerInfo;
-import org.openRealmOfStars.player.leader.LeaderUtility;
+import org.openRealmOfStars.player.leader.LeaderBiography;
 import org.openRealmOfStars.player.leader.Perk;
 
 /**
@@ -158,11 +158,11 @@ public class RealmView extends BlackPanel {
       if (sb.length() > 0) {
         leaderDesc = "<html>" + realm.getRuler().getDescription()
             + "<ht><br>Perks:<br>" + sb.toString()
-            + "<hr><br>" + LeaderUtility.createBioForLeader(realm.getRuler(),
+            + "<hr><br>" + LeaderBiography.createBioForLeader(realm.getRuler(),
                 realm) + "</html>";
       } else {
         leaderDesc = "<html>" + realm.getRuler().getDescription()
-            + "<hr><br>" + LeaderUtility.createBioForLeader(realm.getRuler(),
+            + "<hr><br>" + LeaderBiography.createBioForLeader(realm.getRuler(),
                 realm) + "</html>";
       }
       leaderDesc = leaderDesc.replace("\n", "<br>");
