+++ src/main/java/org/openRealmOfStars/player/leader/LeaderUtility.java
@@ -1055,667 +1055,4 @@ public final class LeaderUtility {
     return false;
   }
 
-  /**
-   * Get Main job description for leader.
-   * @param leader Leader for getting the main job.
-   * @param government Government type for getting proper title for ruler.
-   * @return Job title
-   */
-  private static String getMainJob(final Leader leader,
-      final GovernmentType government) {
-    int ruler = leader.getStats().getStat(StatType.RULER_REIGN_LENGTH);
-    int governor = leader.getStats().getStat(StatType.GOVERNOR_LENGTH);
-    int commander = leader.getStats().getStat(StatType.COMMANDER_LENGTH);
-    int avg = (ruler + governor + commander) / 3;
-    if (ruler > avg) {
-      return RulerUtility.getRulerTitle(leader.getGender(), government);
-    }
-    if (avg == 0) {
-      return "jobless";
-    }
-    if (governor > avg) {
-      return "Governor";
-    }
-    if (commander > avg) {
-      return "Commander";
-    }
-    return "between jobs";
-  }
-
-  /**
-   * Get textual description where leader is known about.
-   * @param leader Leader.
-   * @return String
-   */
-  private static String getBestKnown(final Leader leader) {
-    StringBuilder sb = new StringBuilder();
-    boolean and = false;
-    boolean noMore = false;
-    int numberOfBattle = leader.getStats().getStat(StatType.NUMBER_OF_BATTLES);
-    int numberOfAnomalies = leader.getStats().getStat(
-        StatType.NUMBER_OF_ANOMALY);
-    int trades = leader.getStats().getStat(
-        StatType.NUMBER_OF_TRADES);
-    int privateering = leader.getStats().getStat(
-        StatType.NUMBER_OF_PRIVATEERING);
-    int numberOfPlanets = leader.getStats().getStat(
-        StatType.NUMBER_OF_PLANETS_EXPLORED);
-    int commanderAvg = (numberOfAnomalies + numberOfBattle + trades
-        + privateering + numberOfPlanets) / 5;
-    int numberOfBuildings = leader.getStats().getStat(
-        StatType.NUMBER_OF_BUILDINGS_BUILT);
-    int numberOfShips = leader.getStats().getStat(
-        StatType.NUMBER_OF_SHIPS_BUILT);
-    int populationGrowth = leader.getStats().getStat(
-        StatType.POPULATION_GROWTH);
-    int governorAvg = (numberOfBuildings + numberOfShips
-        + populationGrowth) / 3;
-    int warDeclarations = leader.getStats().getStat(StatType.WAR_DECLARATIONS);
-    int diplomaticTrades = leader.getStats().getStat(StatType.DIPLOMATIC_TRADE);
-    int rulerAvg = (warDeclarations + diplomaticTrades) / 2;
-    if (rulerAvg >= commanderAvg && rulerAvg > 0) {
-      if (warDeclarations > diplomaticTrades) {
-        sb.append("war declarations");
-        and = true;
-      } else {
-        sb.append("diplomatic trades");
-        and = true;
-      }
-    }
-    if (governorAvg >= commanderAvg) {
-      if (numberOfShips > numberOfBuildings
-          && numberOfShips > populationGrowth) {
-        if (and) {
-          noMore = true;
-          sb.append(" and ");
-        }
-        sb.append("ship building");
-        and = true;
-      }
-      if (!noMore && numberOfBuildings > numberOfShips
-          && numberOfBuildings > populationGrowth) {
-        if (and) {
-          noMore = true;
-          sb.append(" and ");
-        }
-        sb.append("building projects");
-        and = true;
-      }
-      if (!noMore && populationGrowth > numberOfShips
-          && populationGrowth > numberOfBuildings) {
-        if (and) {
-          noMore = true;
-          sb.append(" and ");
-        }
-        sb.append("population growth");
-        and = true;
-      }
-    }
-    if (commanderAvg >= governorAvg) {
-      if (!noMore && numberOfBattle > numberOfAnomalies
-          && numberOfBattle > privateering
-          && numberOfBattle > trades
-          && numberOfBattle > numberOfPlanets) {
-        if (and) {
-          noMore = true;
-          sb.append(" and ");
-        }
-        sb.append("space battles");
-        if (leader.getStats().getStat(StatType.NUMBER_OF_PIRATE_BATTLES)
-            >= numberOfBattle / 2) {
-          sb.append(" against space pirates");
-        }
-        and = true;
-      }
-      if (!noMore && numberOfAnomalies > numberOfBattle
-          && numberOfAnomalies > privateering
-          && numberOfAnomalies > trades
-          && numberOfAnomalies > numberOfPlanets) {
-        if (and) {
-          noMore = true;
-          sb.append(" and ");
-        }
-        sb.append("exploring space anomalies");
-        and = true;
-      }
-      if (!noMore && trades > numberOfBattle
-          && trades > privateering
-          && trades > numberOfAnomalies
-          && trades > numberOfPlanets) {
-        if (and) {
-          noMore = true;
-          sb.append(" and ");
-        }
-        sb.append("trades");
-        and = true;
-      }
-      if (!noMore && numberOfPlanets > numberOfBattle
-          && numberOfPlanets > trades
-          && numberOfPlanets > numberOfAnomalies
-          && numberOfPlanets > privateering) {
-        if (and) {
-          noMore = true;
-          sb.append(" and ");
-        }
-        sb.append("boldly exploring planets");
-        and = true;
-      }
-      if (!noMore && privateering > numberOfBattle
-          && privateering > trades
-          && privateering > numberOfAnomalies
-          && privateering > numberOfPlanets) {
-        if (and) {
-          noMore = true;
-          sb.append(" and ");
-        }
-        sb.append("space pirating");
-        and = true;
-      }
-    }
-    return sb.toString();
-  }
-  /**
-   * Create Bio for leader
-   * @param leader Leader whom to create bio.
-   * @param info PlayerInfo
-   * @return Bio as a String.
-   */
-  public static String createBioForLeader(final Leader leader,
-      final PlayerInfo info) {
-    boolean living = true;
-    boolean young = false;
-    if (leader.getAge() < 35) {
-      young = true;
-    }
-    if (leader.getJob() == Job.DEAD) {
-      living = false;
-      young = false;
-    }
-    StringBuilder sb = new StringBuilder();
-    sb.append(leader.getName());
-    if (leader.getJob() == Job.TOO_YOUNG) {
-      sb.append(" is still growing up and will achieve many things later.");
-      return sb.toString();
-    }
-    if (living) {
-      sb.append(" is ");
-    } else {
-      sb.append(" was ");
-    }
-    int ruler = leader.getStats().getStat(StatType.RULER_REIGN_LENGTH);
-    int governor = leader.getStats().getStat(StatType.GOVERNOR_LENGTH);
-    int commander = leader.getStats().getStat(StatType.COMMANDER_LENGTH);
-    if (ruler > 0 && governor == 0 && commander == 0) {
-      sb.append("the ");
-    } else if (ruler == 0 && governor > 0 && commander == 0
-        || ruler == 0 && governor == 0 && commander > 0) {
-      sb.append("working as ");
-    } else if (ruler != 0 || governor != 0 || commander != 0) {
-      sb.append("mostly working as ");
-    }
-    sb.append(getMainJob(leader, info.getGovernment()));
-    if (living) {
-      sb.append(". Currently ");
-      sb.append(leader.getName());
-      sb.append(" is ");
-      if (leader.getJob() == Job.RULER) {
-        sb.append(RulerUtility.getRulerTitle(leader.getGender(),
-            info.getGovernment()));
-      } else if (leader.getJob() == Job.COMMANDER) {
-        sb.append(leader.getMilitaryRank().toString());
-      } else {
-        sb.append(leader.getJob().toString().toLowerCase());
-      }
-      sb.append(". ");
-    } else {
-      sb.append(". ");
-    }
-    sb.append(leader.getTitle());
-    String known = getBestKnown(leader);
-    if (known.isEmpty()) {
-      if (young && living) {
-        sb.append(" is still young and is able to achieve many things. ");
-      }
-      if (!young && living) {
-        sb.append(" is still live and ");
-        if (leader.getRace().isRoboticRace()) {
-          sb.append("functional and");
-        } else if (leader.hasPerk(Perk.HEALTHY)) {
-          sb.append("has healthy lifestyle and");
-        } else if (!leader.hasPerk(Perk.ADDICTED)) {
-          sb.append("healthy and");
-        }
-        sb.append(" is able to achieve things. ");
-      }
-      if (!living) {
-        sb.append(" has passed away with respect. ");
-      }
-    } else {
-      if (living) {
-        sb.append(" is known for ");
-      } else {
-        sb.append(" will be remembered for ");
-      }
-      sb.append(known);
-      sb.append(". ");
-    }
-    if (leader.getStats().getStat(StatType.NUMBER_OF_ESPIONAGE) > 0) {
-      sb.append(leader.getName());
-      if (living) {
-        sb.append(" is suspected to be spy. ");
-      } else {
-        sb.append(" was suspected to be spy. ");
-      }
-    }
-    if (leader.getStats().getStat(StatType.NUMBER_OF_JAIL_TIME) > 0) {
-      sb.append(leader.getName());
-      if (living) {
-        sb.append(" has been in jail. ");
-      } else {
-        sb.append(" has been sentenced to jail ");
-        sb.append(leader.getStats().getStat(StatType.NUMBER_OF_JAIL_TIME));
-        sb.append(" times. ");
-      }
-    }
-    Attitude attitude = getRulerAttitude(leader);
-    if (attitude != null) {
-      sb.append(leader.getCallName());
-      if (living) {
-        sb.append(" is known to be ");
-      } else {
-        sb.append(" was known to be ");
-      }
-      switch (attitude) {
-      case AGGRESSIVE: {
-        sb.append("aggressive");
-        break;
-      }
-      case BACKSTABBING: {
-        sb.append("untrustworthy");
-        break;
-      }
-      case DIPLOMATIC: {
-        sb.append("diplomatic");
-        break;
-      }
-      case EXPANSIONIST: {
-        sb.append("adventurous");
-        break;
-      }
-      case LOGICAL: {
-        sb.append("very logical");
-        break;
-      }
-      case MERCHANTICAL: {
-        sb.append("merchantical");
-        break;
-      }
-      case MILITARISTIC: {
-        sb.append("militaristic");
-        break;
-      }
-      case PEACEFUL: {
-        sb.append("calm and peaceful");
-        break;
-      }
-      default:
-      case SCIENTIFIC: {
-        sb.append("scientific");
-        break;
-      }
-      }
-      sb.append(". ");
-    }
-    if (leader.getStats().getStat(StatType.RESEARCH_ARTIFACTS) > 0) {
-      sb.append(leader.getName());
-      boolean famous = false;
-      if (leader.getStats().getStat(StatType.RESEARCH_ARTIFACTS) > 2) {
-        famous = true;
-      }
-      if (living) {
-        sb.append(" is ");
-      } else {
-        sb.append(" was ");
-      }
-      if (famous) {
-        sb.append("famous ancient artifact researcher. ");
-      } else {
-        sb.append("interested in ancient artifact research. ");
-      }
-    }
-    return sb.toString();
-  }
-
-  /**
-   * Get Reason for gaining perk for leader.
-   * @param leader Leader who is gain perk
-   * @param perk Perk which was gained.
-   * @return Explanation string
-   */
-  public static String getReasonForPerk(final Leader leader, final Perk perk) {
-    StringBuilder sb = new StringBuilder();
-    switch (perk) {
-      default: {
-        sb.append(leader.getName());
-        sb.append(" gained perk called ");
-        sb.append(perk.getName());
-        sb.append(".");
-        break;
-      }
-      case ACADEMIC: {
-        sb.append(leader.getName());
-        sb.append(" has higher level of studies and learn how to be academic.");
-        break;
-      }
-      case ADDICTED: {
-        sb.append(leader.getName());
-        sb.append(" has hooked on substance and ");
-        sb.append(leader.getGender().getHeShe());
-        sb.append(" has to use it.");
-        break;
-      }
-      case AGGRESSIVE: {
-        sb.append(leader.getName());
-        sb.append(" has become aggressive due ");
-        sb.append(leader.getGender().getHisHer());
-        sb.append(" stress levels.");
-        break;
-      }
-      case AGRICULTURAL: {
-        sb.append(leader.getName());
-        sb.append(" has started to grown ");
-        sb.append(leader.getGender().getHisHer());
-        sb.append(" own farm and thus become more interested in agriculture.");
-        break;
-      }
-      case ARCHAEOLOGIST: {
-        sb.append(leader.getName());
-        sb.append(" has researched ancient artifacts and ");
-        sb.append(leader.getGender().getHeShe());
-        sb.append(" has done couple of new findings.");
-        break;
-      }
-      case ARTISTIC: {
-        sb.append(leader.getName());
-        sb.append(" has started to ");
-        switch (DiceGenerator.getRandom(3)) {
-          default:
-          case 0: {
-            sb.append("paint ");
-            break;
-          }
-          case 1: {
-            sb.append("sculture ");
-            break;
-          }
-          case 2: {
-            sb.append("composing music ");
-            break;
-          }
-          case 3: {
-            sb.append("writing novels ");
-            break;
-          }
-        }
-        sb.append("and ");
-        sb.append(leader.getGender().getHisHer());
-        sb.append(" interest towards art has been increased.");
-        break;
-      }
-      case CHARISMATIC: {
-        sb.append(leader.getName());
-        sb.append(" has learn how to behave charismatically.");
-        break;
-      }
-      case CHATTERBOX: {
-        sb.append(leader.getName());
-        sb.append(" feels compulsary need to brag ");
-        sb.append(leader.getGender().getHisHer());
-        sb.append(" achievements and always speaks a too much.");
-        break;
-      }
-      case COMBAT_MASTER: {
-        sb.append(leader.getName());
-        sb.append(" has improved how to shoot accurately. ");
-        break;
-      }
-      case COMBAT_TACTICIAN: {
-        sb.append(leader.getName());
-        sb.append(" has learned how to react with ");
-        sb.append(leader.getGender().getHisHer());
-        sb.append(" fleet in space combat faster than before.");
-        break;
-      }
-      case CONVICT: {
-        sb.append(leader.getName());
-        sb.append(" has been in prison. ");
-        break;
-      }
-      case CORRUPTED: {
-        sb.append(leader.getName());
-        sb.append(" takes realm's credits into ");
-        sb.append(leader.getGender().getHisHer());
-        sb.append(" pockets and ");
-        sb.append(leader.getGender().getHeShe());
-        sb.append(" thinks that is the right thing to do.");
-        break;
-      }
-      case COUNTER_AGENT: {
-        sb.append(leader.getName());
-        sb.append(" has learn how to spot cloaked fleets by noticing little"
-            + " things in scanner views.");
-        break;
-      }
-      case CRUEL: {
-        sb.append(leader.getName());
-        sb.append(" has killed too young heir. What cruel thing to do!");
-        break;
-      }
-      case DIPLOMATIC: {
-        sb.append(leader.getName());
-        sb.append(" has developped skill to be a diplomatic.");
-        break;
-      }
-      case DISCIPLINE: {
-        sb.append(leader.getName());
-        sb.append(" has learned to be strict commander during wartime.");
-        break;
-      }
-      case EXPLORER: {
-        sb.append(leader.getName());
-        sb.append(" is very keen to explore and"
-            + " pushes the engines a bit over the limit.");
-        break;
-      }
-      case FTL_ENGINEER: {
-        sb.append(leader.getName());
-        sb.append(" wants to get moving fast and"
-            + " pushes the engines a bit over the limit.");
-        break;
-      }
-      case GOOD_LEADER: {
-        sb.append(leader.getName());
-        sb.append(" has learned new people skill:"
-            + " how to treat every one equally and fairly.");
-        break;
-      }
-      case HEALTHY: {
-        sb.append(leader.getName());
-        sb.append(" has started to live more healthier life than before.");
-        break;
-      }
-      case INCOMPETENT: {
-        sb.append(leader.getName());
-        sb.append(" has fallen down and hit ");
-        sb.append(leader.getGender().getHisHer());
-        sb.append(" head very badly. Things aren't the same anymore...");
-        break;
-      }
-      case INDUSTRIAL: {
-        sb.append(leader.getName());
-        sb.append(" has shown interest in factory automation ");
-        sb.append(" and seems like ");
-        sb.append(leader.getGender().getHeShe());
-        sb.append(" can get the factories running much more efficiently.");
-        break;
-      }
-      case LOGICAL: {
-        sb.append(leader.getName());
-        sb.append(" has starting to behave logically. This is ");
-        sb.append(leader.getGender().getHisHer());
-        sb.append(" new way to get things done.");
-        break;
-      }
-      case MAD: {
-        sb.append(leader.getName());
-        sb.append(" had very bad disease and ");
-        sb.append(leader.getGender().getHeShe());
-        sb.append(" became mad afterwards.");
-        break;
-      }
-      case MASTER_ENGINEER: {
-        sb.append(leader.getName());
-        sb.append(" has deeper knowledge how to overload ship's"
-            + " components with less danger.");
-        break;
-      }
-      case MERCHANT: {
-        sb.append(leader.getName());
-        sb.append(" has interest in trading and thus makes better deals.");
-        break;
-      }
-      case MICRO_MANAGER: {
-        sb.append(leader.getCallName());
-        sb.append(" has habit that everything needs to be done by ");
-        sb.append(leader.getGender().getHisHer());
-        sb.append(" way and this kind of management does not end well.");
-        break;
-      }
-      case MILITARISTIC: {
-        sb.append(leader.getName());
-        sb.append(" has been interesting in military tactics and can handle"
-            + " larger number of fleets.");
-        break;
-      }
-      case MINER: {
-        sb.append(leader.getName());
-        sb.append(" has been interested in minerals and stone and therefore ");
-        sb.append(leader.getGender().getHeShe());
-        sb.append(" is good spotting great mining areas.");
-        break;
-      }
-      case PACIFIST: {
-        sb.append(leader.getCallName());
-        sb.append(" thinks that war is never a solution and there must be"
-            + " a peaceful way to solve things.");
-        break;
-      }
-      case PEACEFUL: {
-        sb.append(leader.getCallName());
-        sb.append(" has learned how to calm ");
-        sb.append(leader.getGender().getHisHer());
-        sb.append(" nerves. This gives a cool impression.");
-        break;
-      }
-      case POWER_HUNGRY: {
-        sb.append(leader.getCallName());
-        sb.append(" is graving to be a ruler and ");
-        sb.append(leader.getGender().getHeShe());
-        sb.append(" is willing to do lot for it, some of those can be"
-            + " questionable.");
-        break;
-      }
-      case REPULSIVE: {
-        sb.append(leader.getName());
-        sb.append(" has turned into selfish and arrogant that ");
-        sb.append(" doesn't seem to care about others.");
-        break;
-      }
-      case SCANNER_EXPERT: {
-        sb.append(leader.getName());
-        sb.append(" has learned how scanners work and ");
-        sb.append(leader.getGender().getHeShe());
-        sb.append(" knows how to manually tweak in order get them"
-            + " perform better.");
-        break;
-      }
-      case SCIENTIST: {
-        sb.append(leader.getName());
-        sb.append(" has published scientific paper during ");
-        sb.append(leader.getGender().getHisHer());
-        sb.append(" spare time. Impressive.");
-        break;
-      }
-      case SECRET_AGENT: {
-        sb.append(leader.getName());
-        sb.append(" has learned how to tune cloaking ");
-        sb.append(" device to perform slightly better.");
-        break;
-      }
-      case SKILLFUL: {
-        sb.append(leader.getName());
-        sb.append(" has become skillfull, ");
-        sb.append(leader.getGender().getHeShe());
-        sb.append(" is able to pick any task and perform it well.");
-        break;
-      }
-      case SLOW_LEARNER: {
-        sb.append(leader.getName());
-        sb.append(" is feeling that learning new things is a burden and ");
-        sb.append(leader.getGender().getHeShe());
-        sb.append(" hasn't any motivation to improve situation.");
-        break;
-      }
-      case SPY_MASTER: {
-        sb.append(leader.getName());
-        sb.append(" has watched couple of old spy movies and ");
-        sb.append(" learned new tricks for espionage.");
-        break;
-      }
-      case STUPID: {
-        sb.append(leader.getName());
-        sb.append(" has got too many hits to ");
-        sb.append(leader.getGender().getHisHer());
-        sb.append(" head and become simply stupid.");
-        break;
-      }
-      case TRADER: {
-        sb.append(leader.getName());
-        sb.append(" has got interested in trading ");
-        sb.append(" and gains a bit of extra while doing it.");
-        break;
-      }
-      case WARLORD: {
-        sb.append(leader.getName());
-        sb.append(" has learned motivate population during war ");
-        sb.append(" so that they work more like a team.");
-        break;
-      }
-      case WEAK_LEADER: {
-        sb.append(leader.getName());
-        sb.append(" has started to be affraid of war and this is bad ");
-        sb.append(" since fear spreads specially during war times.");
-        break;
-      }
-      case WEALTHY: {
-        sb.append(leader.getName());
-        sb.append(" has got some personal extra credits. ");
-        break;
-      }
-      case TREKKER: {
-        sb.append(leader.getName());
-        sb.append(" has show great interest for exploring uncolonized"
-            + " planets. ");
-        break;
-      }
-      case NEGOTIATOR: {
-        sb.append(leader.getName());
-        sb.append(" had many diplomatic meetings so, ");
-        sb.append(leader.getGender().getHeShe());
-        sb.append(" is able get bit of extra information.");
-        break;
-      }
-    }
-    return sb.toString();
-  }
 }
